requirements.txt
discord.py>=2.4.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.2
